update vote manager to use cache

+++ src/main/java/com/google/step/datamanager/VoteWithExpiry.java
@@ -1,10 +1,10 @@
 package com.google.step.datamanager;
 
 public class VoteWithExpiry {
-  public final long votes;
+  public final int votes;
   public final boolean isExpired;
 
-  public VoteWithExpiry(long votes, boolean isExpired) {
+  public VoteWithExpiry(int votes, boolean isExpired) {
     this.votes = votes;
     this.isExpired = isExpired;
   }
